+++ src/musics/musics.service.ts
@@ -1,4 +1,8 @@
-import { Injectable } from '@nestjs/common';
+import {
+  BadRequestException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { CreateMusicDto } from './dto/create-music.dto';
 import { UpdateMusicDto } from './dto/update-music.dto';
 import { InjectRepository } from '@nestjs/typeorm';
@@ -11,21 +15,53 @@ export class MusicsService {
     @InjectRepository(Music) private musicRepository: Repository<Music>,
   ) {}
 
-  create(createMusicDto: CreateMusicDto) {
+  async create(createMusicDto: CreateMusicDto) {
+    const { name, level, mode } = createMusicDto;
+
+    const identicalMusic = await this.musicRepository.findOne({
+      where: { name: name, level: level, mode: mode },
+    });
+
+    if (identicalMusic) {
+      throw new BadRequestException(
+        'Music with exact same data already exists',
+      );
+    }
+
     const newMusic = this.musicRepository.create({ ...createMusicDto });
-    return this.musicRepository.save(newMusic);
+
+    return await this.musicRepository.save(newMusic);
   }
 
-  findAll() {
-    return this.musicRepository.find();
+  async findAll() {
+    return await this.musicRepository.find({
+      relations: { categories: { event: true } },
+    });
   }
 
-  findOne(id: number) {
-    return this.musicRepository.findOneBy({ music_id: id });
+  async findOne(id: number) {
+    const music = await this.musicRepository.findOne({
+      where: { music_id: id },
+      relations: { categories: { event: true }, scores: { player: true } },
+    });
+
+    if (!music) {
+      throw new NotFoundException('Music not Found');
+    }
+
+    return music;
   }
 
-  update(id: number, updateMusicDto: UpdateMusicDto) {
-    return this.musicRepository.update(
+  async update(id: number, updateMusicDto: UpdateMusicDto) {
+    const music = await this.musicRepository.findOne({
+      where: { music_id: id },
+    });
+
+    if (!music) {
+      throw new NotFoundException('Music not found');
+    }
+
+    return await this.musicRepository.update(
       {
         music_id: id,
       },
@@ -33,7 +69,15 @@ export class MusicsService {
     );
   }
 
-  remove(id: number) {
-    return this.musicRepository.delete({ music_id: id });
+  async remove(id: number) {
+    const music = await this.musicRepository.findOne({
+      where: { music_id: id },
+    });
+
+    if (!music) {
+      throw new NotFoundException('Music not found');
+    }
+
+    return await this.musicRepository.delete({ music_id: id });
   }
 }
